+++ lib/distributions/paper/velocity_distribution.dart
@@ -1,0 +1,15 @@
+import 'paper_distribution.dart';
+
+class VelocityDistribution extends PaperDistribution {
+  @override
+  String get displayName => 'Velocity';
+
+  @override
+  String get project => 'velocity';
+
+  @override
+  bool get hasMetadata => false;
+
+  @override
+  bool get recommended => true;
+}