+++ lib/script/windows_script_generator.dart
@@ -14,6 +14,7 @@ class WindowsScriptGenerator extends ScriptGenerator {
     stringBuffer.writeln('@echo off');
 
     stringBuffer.write(sanitizedJavaPath);
+    stringBuffer.write('/bin/java.exe');
 
     if (additionalArgs.isNotEmpty) {
       stringBuffer.writeln(' ^');