+++ lib/settings/server_chooser.dart
@@ -1,6 +1,5 @@
 import 'package:interact/interact.dart';
 import 'package:mcserv/distributions/distribution.dart';
-import 'package:mcserv/utils/localizations_util.dart';
 import 'package:mcserv/utils/utils.dart';
 
 import 'settings.dart';