+++ CHANGELOG.md
@@ -12,6 +12,33 @@
 ### Fixed
 
 ### Security
+
+## [0.0.6] - 2021-12-29
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+- Fix installer not asking for eula
+- Fix choco CI
+- Fix docs CI
+
+### Security
+
+## [0.0.5] - 2021-12-29
+### Fixed
+- Fix docs CI failing
+- Fix script generators not appending /bin/java
+
+## [0.0.4] - 2021-12-10
+### Changed
+- Update dependencies
+- Make Velocity use the PaperMC API
+
 ## [0.0.3] - 2021-09-08
 ### Fixed
 - Fix Bug causing process to not terminate correctly when downloading items
@@ -19,14 +46,12 @@
 
 ## [0.0.2] - 2021-09-05
 ### Added
-
 - non-interactive CLI flags for all interactive components
 - `--ascii` mode (Use `--no-ascii` on Windows to restore old mode)
 - Loading spinners for expensive tasks (e.g. Unpacking a JDK)
 - Human-readable filesize in Download progress instead of byte count
 
 ### Changed
-
 - Use interact for progress bars
 
 ### Fixed
@@ -35,7 +60,5 @@
 - 64-bit program installed to `Program Files (x86)` on Windows
 
 ## [0.0.1]
-
 ### Added
-
 - Initial version.